requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
hdbscan>=0.8.33
umap-learn>=0.5.4
scipy>=1.10.0
xgboost>=2.0.0
jupyter>=1.0.0
nbformat>=5.9.0
